requirements.txt
langchain
langchain-community
langchain-openai
openai
streamlit
faiss-cpu
pypdf
tiktoken
python-dotenv
huggingface_hub
sentence-transformers
